backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
pydantic>=2.9
pydantic-settings>=2.6
python-multipart>=0.0.9
azure-storage-blob>=12.23
numpy>=1.26
rasterio>=1.3
rio-tiler>=7.4
Pillow>=10.4

